Refactored user table
